Restore the Khronos logo.
Cleanup after recent build mechanism changes.

There is no longer an "inline HTML" version
(we rely on style sheets and mathjax scripts,
so we're not currently single-file anyway).

The Khronos logo had Illustrator proprietary data
in it making it much bigger for no reason; removed.

The output is now in the out/ directory.

Installation scripts have been updated to use
the right files.