Added ASTC blocksize clarification and format fixes
ASTC uses "block footprint" for "texel block size". The latter term
is used in external Khronos standards, but the former is used
exclusively in the ASTC chapter. Added a note that these terms
are synonymous - and added graphical notes so that this was visible.

Fixed the over-sized contributor list and revision log in the PDF.
Restored the contributor list to the HTML (by appending an extra
chapter in HTML only) and removed the redundant text that was
effectively "ifdeffed" out in the new build mechanism. Changed the
spacing of the contents in the PDF so that the numbers don't overlap
the text. Updated the revision number to 1.6.