Fix several typos in BC6H layout tables.
Found while implementing a BC6H decoder from the spec. The new
bit layout tables are generated from the "master" layout table
"table-bptcblockformat", which appears to be correct.

The small Python script to turn the contents of said table into
the layout tables will be included in the next commit.